Fix problems related to max(-0.0, +0.0) returning -0.0 with some languages.
The geodesic routines use the idiom

  atan2(max(y, 0.0), x)

in order to return a result in [0, pi].  However with some languages
with x = -1.0 and y = -0.0 this returns atan2(-0.0, -1.0) = -pi.

This bug isn't exposed in the geodesic routines except for the MATLAB
version running under Octave, where it's triggered by sind(-0.0)
returning +0.0 instead of -0.0.  A test case exposing the bug is

  geoddistance(0,0,0,[179.5,179.6])

returning two negative distances.  And this only became an issue in
version 1.44 and it's only triggered with vector arguments.

Nevertheless, the code needs to be updated to avoid surprises in the
future.

Here is how various languages fare:

* Java: does the right thing, +0.0 is returned.
* C++: well-defined behavior, returning the first argument on equality.
* C: geodesic.c maxx function returns the second argument on equality.
* python, JavaScript, maxima, Fortran: the behavior is undefined.
* MATLAB: also undefined, with the added complication that the behavior
  depends on whether the arguments are vectors or scalars.

The fixes:

* Update the Octave implementation of sincosdx to avoid the use of sind
  and cosd.
* C++: change atan2(max(y, 0.0), x) to atan2(max(0.0, y), x); changes
  needed in Geodesic and GeodesicExact.
* C: this was (accidentally) working OK.  Change definition of maxx to
  mimic the C++ convention and switch the order of the arguments to max.
* Java: this was working OK.  But switch the order of the arguments to
  max to match the C++ convention (a cosmetic change).
* Other languages: use the idiom atan2(0.0 + max(0.0, y), x) to discard
  the sign on -0.0 returned by max.