Fix bug in triaxial.distance; add triaxial.ellipflip
The bug was due to MATLAB's ode solver taking large steps so that it
skipped over the target latitude bet2 and, if bet2 was close to 90, went
over the line bet = 90 so that bet2 started decreasing.  The bug is less
likely with Octave because its ode solver takes smaller steps.  The fix
is to map the pointd "over bet = 90" to the other sheet so that bet2
continues to increase.